+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.2
+- Fixed `route()` such that the single argument is optional
+- Moved `select()` onto `prototype`
+
 ## 1.0.1
 - Created `this.scan()` for reprogramming the router easily